Drawing 3WT in voltage level diagram and substation diagram :
Add an option in the parameters, to show/unshow the self in the tertiary voltage level, when displaying the primary or secondary voltage level diagram.
Displaying the self instead of a winding to the tertiary voltage level, when the show self option is selected.
Colorization of the svg circles/ellipses elements taken from the SVG documents for 2WT and 3WT, in nominal voltage style.
Modification of the SVG for 2WT and 3WT in the FlatDesign library, so this colorization
succeeds.

Signed-off-by: Franck LECUYER <franck.lecuyer@rte-france.com>