Improving 3 windings transformer representation in diagrams :
Changing the SVG document for 3WT.
Modifiying the anchor points for 3WT.
Managing the SVG and anchor points rotation, when 3WT cell is bottom oriented.
Colorizing the wires between the 3WT nodes and the feeder nodes.
Removing the SVG windings for the feeder nodes associated to the 3WT.
Colorizing the inductor SVG.
Adapting the test cases (java, svg, json).

Signed-off-by: Franck LECUYER <franck.lecuyer@rte-france.com>