Fix: better way to import svgdotjs
The previous method broke in another website because the object extended by the
panzoom plugin (added panZoom function to prototype) was not the same as the
object used when calling SVG(): error, no function panZoom() on object. Somehow
when panzoom was extending, the extended object was reported as a plain object
in the console, but when SVG() was called it was the class. Maybe this depends
on the babel version or the webpack version.

Signed-off-by: Jon Harper <jon.harper87@gmail.com>